Fix omit field not registering
The omit field in the default game is normally used for omitting the challenge deck from the center_pools table, however steamodded did not take this into account. I fixed this issue by just checking whether the field is true or not and also adding a new explicit field to the Back constructor.